pyproject.toml
[tool.poetry.scripts]
train = "ganx.cli.train:main"

[tool.poetry]
name = "ganx"
version = "0.1.0"
description = ""
authors = ["Ross Hemsley <rlhemsley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
jax = "^0.2.9"
jaxlib = "^0.1.59"
dm-haiku = {git = "https://github.com/deepmind/dm-haiku"}
omegaconf = "^2.0.6"
click = "^7.1.2"
chex = "^0.0.3"
pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
